+++ tests/test_describe_model.py
@@ -16,7 +16,12 @@ from tests.testmodels import (
 )
 from tortoise import Tortoise, fields
 from tortoise.contrib import test
-from tortoise.fields.relational import BackwardFKRelation
+from tortoise.fields.relational import (
+    BackwardFKRelation,
+    ForeignKeyFieldInstance,
+    ManyToManyFieldInstance,
+    OneToOneFieldInstance,
+)
 
 
 class TestBasic(test.TestCase):
@@ -283,7 +288,7 @@ class TestBasic(test.TestCase):
                 "fk_fields": [
                     {
                         "name": "fk",
-                        "field_type": fields.ForeignKeyFieldInstance,
+                        "field_type": ForeignKeyFieldInstance,
                         "raw_field": "fk_id",
                         "python_type": StraightFields,
                         "generated": False,
@@ -311,7 +316,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
-                        "field_type": fields.OneToOneFieldInstance,
+                        "field_type": OneToOneFieldInstance,
                         "generated": False,
                         "indexed": True,
                         "name": "o2o",
@@ -337,7 +342,7 @@ class TestBasic(test.TestCase):
                 "m2m_fields": [
                     {
                         "name": "rel_to",
-                        "field_type": fields.ManyToManyFieldInstance,
+                        "field_type": ManyToManyFieldInstance,
                         "python_type": StraightFields,
                         "generated": False,
                         "nullable": False,
@@ -348,7 +353,7 @@ class TestBasic(test.TestCase):
                     },
                     {
                         "name": "rel_from",
-                        "field_type": fields.ManyToManyFieldInstance,
+                        "field_type": ManyToManyFieldInstance,
                         "python_type": StraightFields,
                         "generated": False,
                         "nullable": False,
@@ -603,7 +608,7 @@ class TestBasic(test.TestCase):
                 "fk_fields": [
                     {
                         "name": "fk",
-                        "field_type": fields.ForeignKeyFieldInstance,
+                        "field_type": ForeignKeyFieldInstance,
                         "raw_field": "fk_id",
                         "python_type": SourceFields,
                         "generated": False,
@@ -631,7 +636,7 @@ class TestBasic(test.TestCase):
                     {
                         "default": None,
                         "description": "Line",
-                        "field_type": fields.OneToOneFieldInstance,
+                        "field_type": OneToOneFieldInstance,
                         "generated": False,
                         "indexed": True,
                         "name": "o2o",
@@ -657,7 +662,7 @@ class TestBasic(test.TestCase):
                 "m2m_fields": [
                     {
                         "name": "rel_to",
-                        "field_type": fields.ManyToManyFieldInstance,
+                        "field_type": ManyToManyFieldInstance,
                         "python_type": SourceFields,
                         "generated": False,
                         "nullable": False,
@@ -668,7 +673,7 @@ class TestBasic(test.TestCase):
                     },
                     {
                         "name": "rel_from",
-                        "field_type": fields.ManyToManyFieldInstance,
+                        "field_type": ManyToManyFieldInstance,
                         "python_type": SourceFields,
                         "generated": False,
                         "nullable": False,
@@ -806,7 +811,7 @@ class TestBasic(test.TestCase):
                 "m2m_fields": [
                     {
                         "name": "peers",
-                        "field_type": fields.ManyToManyFieldInstance,
+                        "field_type": ManyToManyFieldInstance,
                         "python_type": UUIDM2MRelatedModel,
                         "generated": False,
                         "nullable": False,