+++ tests/test_postgres.py
@@ -56,7 +56,7 @@ class TestTwoDatabases(test.SimpleTestCase):
         self.db_config["connections"]["models"]["credentials"]["ssl"] = True
         try:
             await Tortoise.init(self.db_config)
-        except (ConnectionError, ssl.SSLCertVerificationError):
+        except (ConnectionError, ssl.SSLError):
             pass
         else:
             self.assertFalse(True, "Expected ConnectionError or SSLCertVerificationError")